# src/main/java/com/sabrina/librarymanagementsystem/controller/api/AdminApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.15.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sabrina.librarymanagementsystem.controller.api;


import com.sabrina.librarymanagementsystem.controller.dto.UserResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-25T22:50:01.095706157+02:00[Europe/Andorra]", comments = "Generator version: 7.15.0")
@Validated
@Tag(name = "Admin", description = "Administration management")
public interface AdminApi {

    public static final String PATH_API_ADMIN_USERS_ID_MAKE_ADMIN_PUT = "/api/admin/users/{id}/make-admin";
    /**
     * PUT /api/admin/users/{id}/make-admin : Promote user to admin
     * Change the role of a user to ADMIN.
     *
     * @param id User id (required)
     * @return User promoted successfully (status code 200)
     *         or User not found (status code 404)
     *         or Access denied (status code 403)
     */
    @Operation(
        operationId = "apiAdminUsersIdMakeAdminPut",
        summary = "Promote user to admin",
        description = "Change the role of a user to ADMIN.",
        tags = { "Admin" },
        responses = {
            @ApiResponse(responseCode = "200", description = "User promoted successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "User not found"),
            @ApiResponse(responseCode = "403", description = "Access denied")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = AdminApi.PATH_API_ADMIN_USERS_ID_MAKE_ADMIN_PUT,
        produces = { "application/json" }
    )
    
    ResponseEntity<UserResponse> apiAdminUsersIdMakeAdminPut(
        @Parameter(name = "id", description = "User id", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );

}