# src/main/java/com/sabrina/librarymanagementsystem/controller/api/AuthenticationApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.15.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sabrina.librarymanagementsystem.controller.api;



import com.sabrina.librarymanagementsystem.controller.dto.AuthenticationResponse;
import com.sabrina.librarymanagementsystem.controller.dto.LoginRequest;
import com.sabrina.librarymanagementsystem.controller.dto.RegisterRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

import jakarta.validation.Valid;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-24T21:46:52.751498263+02:00[Europe/Andorra]", comments = "Generator version: 7.15.0")
@Validated
@Tag(name = "Authentication", description = "Authentication management")
public interface AuthenticationApi {

    public static final String PATH_API_AUTH_LOGIN_POST = "/api/auth/login";
    /**
     * POST /api/auth/login : Authenticate a user
     * Authenticates a user and returns a JWT token.
     *
     * @param loginRequest  (required)
     * @return Authentication successful (status code 200)
     *         or Invalid request (status code 400)
     *         or Invalid email or password (status code 401)
     */
    @Operation(
        operationId = "apiAuthLoginPost",
        summary = "Authenticate a user",
        description = "Authenticates a user and returns a JWT token.",
        tags = { "Authentication" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Authentication successful", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthenticationResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request"),
            @ApiResponse(responseCode = "401", description = "Invalid email or password")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = AuthenticationApi.PATH_API_AUTH_LOGIN_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<AuthenticationResponse> apiAuthLoginPost(
        @Parameter(name = "LoginRequest", description = "", required = true) @Valid @RequestBody LoginRequest loginRequest
    );


    public static final String PATH_API_AUTH_REGISTER_POST = "/api/auth/register";
    /**
     * POST /api/auth/register : Register a new user
     * Creates a new user account.
     *
     * @param registerRequest  (required)
     * @return User successfully registered (status code 201)
     *         or Invalid request (status code 400)
     *         or Email already exists (status code 409)
     */
    @Operation(
        operationId = "apiAuthRegisterPost",
        summary = "Register a new user",
        description = "Creates a new user account.",
        tags = { "Authentication" },
        responses = {
            @ApiResponse(responseCode = "201", description = "User successfully registered", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthenticationResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request"),
            @ApiResponse(responseCode = "409", description = "Email already exists")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = AuthenticationApi.PATH_API_AUTH_REGISTER_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<AuthenticationResponse> apiAuthRegisterPost(
        @Parameter(name = "RegisterRequest", description = "", required = true) @Valid @RequestBody RegisterRequest registerRequest
    );

}