# src/main/java/com/sabrina/librarymanagementsystem/controller/api/AuthorsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.15.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.sabrina.librarymanagementsystem.controller.api;


import com.sabrina.librarymanagementsystem.controller.dto.AuthorRequest;
import com.sabrina.librarymanagementsystem.controller.dto.AuthorResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import java.util.List;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-27T12:11:03.014051287+02:00[Europe/Andorra]", comments = "Generator version: 7.15.0")
@Validated
@Tag(name = "Authors", description = "Author management")
public interface AuthorsApi {

    public static final String PATH_API_AUTHORS_GET = "/api/authors";
    /**
     * GET /api/authors : Get all authors
     *
     * @return List of authors (status code 200)
     */
    @Operation(
        operationId = "apiAuthorsGet",
        summary = "Get all authors",
        tags = { "Authors" },
        responses = {
            @ApiResponse(responseCode = "200", description = "List of authors", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = AuthorResponse.class)))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = AuthorsApi.PATH_API_AUTHORS_GET,
        produces = { "application/json" }
    )
    
    ResponseEntity<List<AuthorResponse>> apiAuthorsGet(
        
    );


    public static final String PATH_API_AUTHORS_ID_DELETE = "/api/authors/{id}";
    /**
     * DELETE /api/authors/{id} : Delete an author
     *
     * @param id  (required)
     * @return Author deleted (status code 204)
     *         or Author not found (status code 404)
     */
    @Operation(
        operationId = "apiAuthorsIdDelete",
        summary = "Delete an author",
        tags = { "Authors" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Author deleted"),
            @ApiResponse(responseCode = "404", description = "Author not found")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = AuthorsApi.PATH_API_AUTHORS_ID_DELETE
    )
    
    ResponseEntity<Void> apiAuthorsIdDelete(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    public static final String PATH_API_AUTHORS_ID_GET = "/api/authors/{id}";
    /**
     * GET /api/authors/{id} : Get author by id
     *
     * @param id  (required)
     * @return Author found (status code 200)
     *         or Author not found (status code 404)
     */
    @Operation(
        operationId = "apiAuthorsIdGet",
        summary = "Get author by id",
        tags = { "Authors" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Author found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Author not found")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = AuthorsApi.PATH_API_AUTHORS_ID_GET,
        produces = { "application/json" }
    )
    
    ResponseEntity<AuthorResponse> apiAuthorsIdGet(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    );


    public static final String PATH_API_AUTHORS_ID_PUT = "/api/authors/{id}";
    /**
     * PUT /api/authors/{id} : Update an author
     *
     * @param id  (required)
     * @param authorRequest  (required)
     * @return Author updated (status code 200)
     *         or Author not found (status code 404)
     */
    @Operation(
        operationId = "apiAuthorsIdPut",
        summary = "Update an author",
        tags = { "Authors" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Author updated", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Author not found")
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = AuthorsApi.PATH_API_AUTHORS_ID_PUT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<AuthorResponse> apiAuthorsIdPut(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "AuthorRequest", description = "", required = true) @Valid @RequestBody AuthorRequest authorRequest
    );


    public static final String PATH_API_AUTHORS_POST = "/api/authors";
    /**
     * POST /api/authors : Create a new author
     *
     * @param authorRequest  (required)
     * @return Author created (status code 201)
     */
    @Operation(
        operationId = "apiAuthorsPost",
        summary = "Create a new author",
        tags = { "Authors" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Author created", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = AuthorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = AuthorsApi.PATH_API_AUTHORS_POST,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<AuthorResponse> apiAuthorsPost(
        @Parameter(name = "AuthorRequest", description = "", required = true) @Valid @RequestBody AuthorRequest authorRequest
    );

}