# src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Library Management API
  description: REST API for managing a library.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server
...
security:
  - bearerAuth: []
paths:
  /api/auth/register:
    post:
      security: []
      tags:
        - Authentication
      summary: Register a new user
      description: Creates a new user account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'

        '400':
          description: Invalid request

        '409':
          description: Email already exists
  /api/auth/login:
    post:
      security: [ ]
      tags:
        - Authentication
      summary: Authenticate a user
      description: Authenticates a user and returns a JWT token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponse'

        '400':
          description: Invalid request

        '401':
          description: Invalid email or password
  /api/admin/users/{id}/make-admin:
    put:
      security:
        - bearerAuth: [ ]
      tags:
        - Admin
      summary: Promote user to admin
      description: Change the role of a user to ADMIN.
      parameters:
        - name: id
          in: path
          required: true
          description: User id
          schema:
            type: integer
            format: int64

      responses:
        '200':
          description: User promoted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

        '404':
          description: User not found

        '403':
          description: Access denied
...
  /api/authors:
    get:
      security:
        - bearerAuth: [ ]

      tags:
        - Authors
      summary: Get all authors
      responses:
        '200':
          description: List of authors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuthorResponse'

    post:
      security:
        - bearerAuth: [ ]

      tags:
        - Authors
      summary: Create a new author
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorRequest'
      responses:
        '201':
          description: Author created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
  /api/authors/{id}:
    get:
      security:
        - bearerAuth: [ ]

      tags:
        - Authors
      summary: Get author by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Author found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
        '404':
          description: Author not found

    put:
      security:
        - bearerAuth: [ ]

      tags:
        - Authors
      summary: Update an author
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthorRequest'
      responses:
        '200':
          description: Author updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthorResponse'
        '404':
          description: Author not found

    delete:
      security:
        - bearerAuth: [ ]

      tags:
        - Authors
      summary: Delete an author
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Author deleted
        '404':
          description: Author not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required:
        - firstname
        - lastname
        - email
        - password

      properties:
        firstname:
          type: string
          minLength: 2
          maxLength: 50
          example: Sabrina

        lastname:
          type: string
          minLength: 2
          maxLength: 50
          example: Moufok

        email:
          type: string
          format: email
          maxLength: 100
          example: sabrina@gmail.com

        password:
          type: string
          minLength: 8
          maxLength: 100
          format: password
          example: password123
    AuthenticationResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
          example: eyJhbGciOiJIUzI1NiJ9...
    LoginRequest:
      type: object
      required:
        - email
        - password

      properties:
        email:
          type: string
          format: email
          maxLength: 100
          example: sabrina@gmail.com

        password:
          type: string
          minLength: 8
          maxLength: 100
          format: password
          example: password123
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64

        firstname:
          type: string
          example: Sabrina

        lastname:
          type: string
          example: Moufok

        email:
          type: string
          example: sabrina@gmail.com

        role:
          type: string
          example: ADMIN
...
    AuthorRequest:
      type: object
      required:
        - firstname
        - lastname

      properties:
        firstname:
          type: string
          minLength: 2
          maxLength: 50
          example: Robert

        lastname:
          type: string
          minLength: 2
          maxLength: 50
          example: Martin

        biography:
          type: string
          maxLength: 1000
          example: Software engineer and author of Clean Code
    AuthorResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64

        firstname:
          type: string
          example: Robert

        lastname:
          type: string
          example: Martin

        biography:
          type: string
          example: Software engineer and author of Clean Code
...